backend/requirements.txt
fastapi>=0.115.0
slowapi>=0.1.9
uvicorn>=0.34.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
websockets>=14.0
redis
numpy
openpyxl
pandas
python-dateutil
six
ecdsa
pyasn1<0.5.0,>=0.4.1
rsa
click
colorama
blinker
Flask
Flask-Login
Flask-SQLAlchemy
itsdangerous
Jinja2
MarkupSafe
Werkzeug
